Le theme persist a la fermeture et ouverture de l'appli
